refactor: switch width/height and thread_count to u32, drop owned_option_keys guard
- change width/height params from usize to u32 across MediaFrame, EncoderBuilder,
  DecoderBuilder, Filter and all examples/tests/benches call sites
- change thread_count from usize to u32 with num_cpus cast on default
- remove owned_option_keys conflict guards and ensure_single_source helper
  (avcodec_open2 dict silently overwrites typed AVCodecContext fields)
- move hw_cache_test_lock() from hwaccel.rs into #[cfg(test)] test_support.rs
- fix leftover 'z z z z' syntax error in frame.rs format! string

+++ examples/color_professional.rs
@@ -19,15 +19,15 @@ fn main() -> anyhow::Result<()> {
     );
 
     // 2. Explicit RGB -> YUV color matrices
-    const W: usize = 320;
-    const H: usize = 180;
+    const W: u32 = 320;
+    const H: u32 = 180;
     let mut rgb = MediaFrame::<u8>::new_video_frame(W, H, PixelFormat::RGB24)?;
     let samples = rgb
         .data
         .as_packed_mut()
         .expect("RGB24 frames are interleaved");
-    for y in 0..H {
-        for x in 0..W {
+    for y in 0..H as usize {
+        for x in 0..W as usize {
             let t = (x as f32 / W as f32 * 255.0) as u8;
             samples[[y, x, 0]] = t;
             samples[[y, x, 1]] = 128;
